nav bar, search form

+++ app/assets/javascripts/festival.js
@@ -1,29 +1,56 @@
 $(function() {
 
+
   // ADD OR REMOVE FESTIVALS FROM FAVORITES ON FESTIVAL SHOW PAGE
-  $('.cache-btns').on('click', '.fave-btn',function() {
-    var driving = $('.driving-cost');
-    var tripCost = {
-      festivalId: $('.music-festival').attr('data-id'),
-      drivingPrice: driving.attr('data-car-price'),
-      drivingTime: driving.attr('data-car-time')
-    };
-    $.ajax('/festival-select',
-      { dataType: 'json',
-        type: 'POST',
-        data: tripCost
-      });
-    $(this).replaceWith('<button class="remove-btn">Remove from Favourites</button>');
+
+  $('.cache-btns').on('change', '.fave-btn', function(){
+    var checkbox = $('#click');
+
+      if (checkbox.prop("checked") == true){
+        var driving = $('.driving-cost');
+        var tripCost = {
+          festivalId: $('.music-festival').attr('data-id'),
+          drivingPrice: driving.attr('data-car-price'),
+          drivingTime: driving.attr('data-car-time')
+        };
+        $.ajax('/festival-select',
+          { dataType: 'json',
+            type: 'POST',
+            data: tripCost
+          });
+      }
+      else if (checkbox.prop("checked") == false){
+        $.ajax('/festival-unselect',
+        { dataType: 'json',
+          type: 'DELETE', 
+          data: { festivalId: $('.music-festival').attr('data-id') }
+        });
+      }
   });
 
-  $('.cache-btns').on('click', '.remove-btn', function() {
-    $(this).replaceWith('<button class="fave-btn">Add to Favourites</button>');
-    $.ajax('/festival-unselect',
-      { dataType: 'json',
-        type: 'DELETE', 
-        data: { festivalId: $('.music-festival').attr('data-id') }
-      });
-  });
+  // $('.cache-btns').on('click', '.fave-btn',function() {
+  //   var driving = $('.driving-cost');
+  //   var tripCost = {
+  //     festivalId: $('.music-festival').attr('data-id'),
+  //     drivingPrice: driving.attr('data-car-price'),
+  //     drivingTime: driving.attr('data-car-time')
+  //   };
+  //   $.ajax('/festival-select',
+  //     { dataType: 'json',
+  //       type: 'POST',
+  //       data: tripCost
+  //     });
+  //   $(this).replaceWith('<button class="remove-btn">Remove from Favourites</button>');
+  // });
+
+  // $('.cache-btns').on('click', '.remove-btn', function() {
+  //   $(this).replaceWith('<button class="fave-btn">Add to Favourites</button>');
+  //   $.ajax('/festival-unselect',
+  //     { dataType: 'json',
+  //       type: 'DELETE', 
+  //       data: { festivalId: $('.music-festival').attr('data-id') }
+  //     });
+  // });
   // REMOVE FESTIVALS FROM FESTIVAL FAVOURITES PAGE
   $('.remove-fave').on('click', function() {
     var festivalDiv = $(this).closest('.fave-festival');
@@ -36,12 +63,69 @@ $(function() {
       });
   });
 
-  debugger;
-
-  var carPrice = $('.driving-cost')[0].dataset.carPrice
-
   //Load Driving Directions on Festival Details page
 
+   
+  var festivalMaps = {
+
+  loadDrivingMap: function() {
+
+    map = new google.maps.Map(document.getElementById('driving-map'), {
+      center: departure,
+      scrollwheel: false,
+      zoom: 8
+    });
+
+    var directionsDisplay = new google.maps.DirectionsRenderer({
+      map: map
+    });
+
+    // Set destination, origin and travel mode.
+    var request = {
+      destination: destination,
+      origin: departure,
+      travelMode: google.maps.TravelMode.DRIVING
+    };
+
+    // Pass the directions request to the directions service.
+    var directionsService = new google.maps.DirectionsService();
+    directionsService.route(request, function(response, status) {
+      if (status == google.maps.DirectionsStatus.OK) {
+        // Display the route on the map.
+        directionsDisplay.setDirections(response);
+        }
+     });
+   },
+  loadFestivalMap: function() {
+
+    map2 = new google.maps.Map(document.getElementById('festival-map'), {
+      center: destination,
+      scrollwheel: false,
+      zoom: 8
+    });
+
+    festivalMarker = new google.maps.Marker({
+      position: destination,
+      map: map2
+   });
+
+  },
+
+  resetDrivingMap: function() {
+    google.maps.event.trigger(map, "resize");
+    map.setCenter(departure);
+    map.setZoom(4);
+    },
+  };
+
+  if ($('#festival-show').length > 0 && carPrice > 0) {
+    // festivalMaps.loadDrivingMap();
+
+  }
+
+  if ($('#festival-show').length > 0) {
+
+    var carPrice = $('.driving-cost')[0].dataset.carPrice
     var drivingMapDiv = $('#travel-tabs').find('#driving-map');
     var destinationCoords = {
       lat: drivingMapDiv[0].dataset.latitude,
@@ -55,64 +139,7 @@ $(function() {
     var map2;
     var festivalMarker;
 
-    var festivalMaps = {
-
-    loadDrivingMap: function() {
-
-      map = new google.maps.Map(document.getElementById('driving-map'), {
-        center: departure,
-        scrollwheel: false,
-        zoom: 8
-      });
-
-      var directionsDisplay = new google.maps.DirectionsRenderer({
-        map: map
-      });
-
-      // Set destination, origin and travel mode.
-      var request = {
-        destination: destination,
-        origin: departure,
-        travelMode: google.maps.TravelMode.DRIVING
-      };
-
-      // Pass the directions request to the directions service.
-      var directionsService = new google.maps.DirectionsService();
-      directionsService.route(request, function(response, status) {
-        if (status == google.maps.DirectionsStatus.OK) {
-          // Display the route on the map.
-          directionsDisplay.setDirections(response);
-          }
-       });
-     },
-    loadFestivalMap: function() {
-
-      map2 = new google.maps.Map(document.getElementById('festival-map'), {
-        center: destination,
-        scrollwheel: false,
-        zoom: 8
-      });
-
-      debugger;
-
-      festivalMarker = new google.maps.Marker({
-        position: destination,
-        map: map2
-     });
-
-  },
-
-  resetDrivingMap: function() {
-    google.maps.event.trigger(map, "resize");
-    map.setCenter(departure);
-    map.setZoom(4);
-    },
-  };
-
-  festivalMaps.loadFestivalMap();
-
-  if ($('#festival-show').length > 0 && carPrice > 0) {
-    festivalMaps.loadDrivingMap();
+    // festivalMaps.loadFestivalMap();
 
   }
   
@@ -134,7 +161,7 @@ $(function() {
 
     if (carPrice > 0) {
 
-      festivalMaps.resetDrivingMap();
+      // festivalMaps.resetDrivingMap();
     }
 
   });
@@ -154,7 +181,7 @@ $(function() {
 
     if (carPrice > 0) {
 
-      festivalMaps.resetDrivingMap();
+      // festivalMaps.resetDrivingMap();
     }
 
   });
@@ -202,8 +229,6 @@ $(function() {
   
 });
 
-
-  console.log("outside function");
   function initMap() {
     var myLatLng = {lat: 49.2827, lng: -123.1207};
     var map;
